Changed Interface and Abstract to double type

+++ AbstractClasses/TwoDimensionalShape.cs
@@ -8,7 +8,7 @@ namespace ShapeCalculatorWPF
 {
     public abstract class TwoDimensionalShape
     {
-        public abstract float Area { get; }
+        public abstract double Area { get; }
 
     }
 }
